+++ tests/examples/amend_delete_both/main.sh
@@ -40,7 +40,7 @@ stepup join
 
 # Wait for background processes, if any.
 set +e; wait -fn $PID; RETURNCODE=$?; set -e
-[[ "${RETURNCODE}" -eq 3 ]] || exit 1
+[[ "${RETURNCODE}" -eq 4 ]] || exit 1
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1