+++ tests/examples/amend_missing/main.sh
@@ -15,7 +15,7 @@ stepup join
 
 # Wait for background processes, if any.
 set +e; wait -fn $PID; RETURNCODE=$?; set -e
-[[ "${RETURNCODE}" -eq 2 ]] || exit 1
+[[ "${RETURNCODE}" -eq 6 ]] || exit 1
 
 # Check files that are expected to be present and/or missing.
 [[ -f plan.py ]] || exit 1