+++ tests/examples/runnable/main.sh
@@ -5,20 +5,20 @@ trap 'kill $(pgrep -g $$ | grep -v $$) > /dev/null 2> /dev/null || :' EXIT
 rm -rvf $(cat .gitignore)
 
 # Run the example
-echo lingering > pong.txt
 stepup boot -n 1 -w & # > current_stdout.txt &
 PID=$!
+
+# Interrupt StepUp with a shutdown command.
+sleep 2
+stepup shutdown
 
 # Get the graph after completion of the pending steps.
 stepup wait
 stepup graph current_graph
-stepup join
-
-# Check files that are expected to be present and/or missing.
-[[ -f plan.py ]] || exit 1
-[[ -f ping.txt ]] || exit 1
-[[ -f pong.txt ]] || exit 1
 
 # Wait for background processes, if any.
 set +e; wait -fn $PID; RETURNCODE=$?; set -e
-[[ "${RETURNCODE}" -eq 4 ]] || exit 1
+[[ "${RETURNCODE}" -eq 12 ]] || exit 1
+
+# Check files that are expected to be present and/or missing.
+[[ -f plan.py ]] || exit 1