+++ tests/test_examples.py
@@ -154,6 +154,7 @@ OVERWRITE_EXPECTED = "STEPUP_OVERWRITE_EXPECTED" in os.environ
         "restart_static_remove1",
         "restart_static_remove2",
         "reqdir_missing",
+        "runnable",
         "script_cases",
         "script_cases_run_import",
         "script_cases_settings",